Removed MainActivityView.kt
