+++ app/build.gradle.kts
@@ -51,11 +51,13 @@ dependencies {
     implementation(project(":lib"))
     implementation(project(":common"))
     implementation("androidx.activity:activity-compose:1.5.0")
+//    implementation(compose.preview)
+    implementation("androidx.compose.ui:ui-tooling-preview:1.2.1")
 
 //    implementation "androidx.core:core-ktx:1.7.0"
 //    implementation "androidx.compose.ui:ui:$compose_version"
 //    implementation "androidx.compose.material:material:$compose_version"
-//    implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
+//    implementation("androidx.compose.ui:ui-tooling-preview:${extra["compose.version"]}")
 //    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
 //    implementation 'androidx.appcompat:appcompat:1.3.0'
 //    implementation 'com.google.android.material:material:1.4.0'
