+++ app/src/main/java/com/github/kentvu/t9vietnamese/android/MainActivity.kt
@@ -4,10 +4,8 @@ import App
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.runtime.*
+import androidx.compose.ui.tooling.preview.Preview
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,4 +15,10 @@ class MainActivity : ComponentActivity() {
             App()
         }
     }
+}
+
+@Preview
+@Composable
+fun PreviewKeyboard() {
+    App()
 }
